refactor: create handlers package
